NoteTest: restructure into header and source file
